requirements.txt
mistralai>=1.0.0
pathlib2>=2.3.0; python_version < '3.4'
python-docx>=1.1.0
python-pptx>=0.6.21
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
jinja2>=3.1.4
